+++ R/axe-generics.R
@@ -22,6 +22,8 @@ butcher <- function(x, verbose = FALSE, ...) {
   x <- axe_data(x, verbose = FALSE, ...)
   x <- axe_env(x, verbose = FALSE, ...)
   x <- axe_fitted(x, verbose = FALSE, ...)
+  x <- axe_rsample_data(x, verbose = FALSE, ...)
+  x <- axe_rsample_indicators(x, verbose = FALSE, ...)
 
   add_butcher_attributes(
     x,
@@ -106,7 +108,6 @@ axe_data.default <- function(x, verbose = FALSE, ...) {
   x
 }
 
-
 #' Axe an environment.
 #'
 #' Remove the environment(s) attached to modeling objects as they are
@@ -159,3 +160,82 @@ axe_fitted.default <- function(x, verbose = FALSE, ...) {
   x
 }
 
+#' Axe data within rsample objects.
+#'
+#' Replace the splitting and resampling objects with a placeholder.
+#'
+#' Resampling and splitting objects produced by \pkg{rsample} contain `rsplit`
+#' objects. These contain the original data set. These data might be large so
+#' we sometimes wish to remove them when saving objects. This method creates a
+#' zero-row slice of the dataset, retaining only the column names and their
+#' attributes, while replacing the original data.
+#'
+#' @name axe-rsample-data
+#' @inheritParams butcher
+#' @param x An object.
+#'
+#' @return An updated object without data in the `rsplit` objects.
+#'
+#' @section Methods:
+#' \Sexpr[stage=render,results=rd]{butcher:::methods_rd("axe_rsample_data")}
+#'
+#' @examplesIf rlang::is_installed("rsample")
+#'
+#' large_cars <- mtcars[rep(1:32, 50), ]
+#' large_cars_split <- rsample::initial_split(large_cars)
+#' butcher(large_cars_split, verbose = TRUE)
+#'
+#' @export
+axe_rsample_data <- function(x, verbose = FALSE, ...) {
+  UseMethod("axe_rsample_data")
+}
+
+#' @export
+#' @name axe-rsample-data
+axe_rsample_data.default <- function(x, verbose = FALSE, ...) {
+  old <- x
+  if (verbose) {
+    assess_object(old, x)
+  }
+  x
+}
+
+#' Axe indicators within rsample objects.
+#'
+#' Replace the splitting and resampling objects with a placeholder.
+#'
+#' Resampling and splitting objects produced by \pkg{rsample} contain `rsplit`
+#' objects. These contain the original data set as well as indicators that
+#' specify which rows go into which data partitions. These size of these
+#' integers might be large so we sometimes wish to remove them when saving
+#' objects. This method saves a zero-row integer in their place.
+#'
+#' @name axe-rsample-indicators
+#' @inheritParams butcher
+#' @param x An object.
+#'
+#' @return An updated object without the indicators in the `rsplit` objects.
+#'
+#' @section Methods:
+#' \Sexpr[stage=render,results=rd]{butcher:::methods_rd("axe_rsample_indicators")}
+#'
+#' @examplesIf rlang::is_installed("rsample")
+#'
+#' large_cars <- mtcars[rep(1:32, 50), ]
+#' large_cars_split <- rsample::initial_split(large_cars)
+#' butcher(large_cars_split, verbose = TRUE)
+#'
+#' @export
+axe_rsample_indicators <- function(x, verbose = FALSE, ...) {
+  UseMethod("axe_rsample_indicators")
+}
+
+#' @export
+#' @name axe-rsample-indicators
+axe_rsample_indicators.default <- function(x, verbose = FALSE, ...) {
+  old <- x
+  if (verbose) {
+    assess_object(old, x)
+  }
+  x
+}