+++ NEWS.md
@@ -2,11 +2,12 @@
 
 * Added methods for `MASS::polr` (@pbulsink, #289).
 
+* Added methods for `rset`, `rsplit`, `tune_results`, and `workflow_set` classes (#292).
+
 * Make to work with new versions of xgboost models (#294).
 
 * Added butcher methods for `tabnet()` (@cregouby #226).
 
-* * Added methods for `MASS::polr` (@pbulsink, #289).
 
 # butcher 0.3.6